remove old test collection (#21513)
