+++ .circleci/config.yml
@@ -259,7 +259,6 @@ references:
       command: |
         python3 -m pytest ./Tests/scripts/infrastructure_tests/ -v
         python3 -m pytest ./Tests/Marketplace/Tests/ -v
-        python3 -m pytest ./Tests/scripts/utils/tests -v
         python3 -m pytest ./Tests/tests -v
         python3 -m pytest ./Tests/private_build/tests -v
         python3 -m pytest Utils -v
