+++ .github/CODEOWNERS
@@ -20,7 +20,7 @@
 /Utils/trigger_upload_packs_to_production.sh @Noy-Maimon
 
 # Test Collection
-/Tests/scripts/collect_tests_and_content_packs.py @dorschw
+/Tests/scripts/collect_tests @dorschw
 
 # PANW Products
 /Packs/Palo_Alto_Networks_Enterprise_DLP/ @DeanArbel
