+++ .gitlab/ci/global.yml
@@ -203,7 +203,6 @@
     - section_start "Test Infrastructure"
     - python3 -m pytest ./Tests/scripts/infrastructure_tests/ -v
     - python3 -m pytest ./Tests/Marketplace/Tests/ -v
-    - python3 -m pytest ./Tests/scripts/utils/tests -v
     - python3 -m pytest ./Tests/tests -v
     - python3 -m pytest ./Tests/private_build/ -v
     - python3 -m pytest Utils -v
