+++ Tests/Marketplace/marketplace_constants.py
@@ -2,8 +2,8 @@ import os
 import json
 import enum
 from typing import List
-from Tests.scripts.utils.content_packs_util import IGNORED_FILES
 
+IGNORED_FILES = ['__init__.py', 'ApiModules', 'NonSupported']  # files to ignore inside Packs folder
 CONTENT_ROOT_PATH = os.path.abspath(os.path.join(__file__, '../../..'))  # full path to content root repo
 PACKS_FOLDER = "Packs"  # name of base packs folder inside content repo
 PACKS_FULL_PATH = os.path.join(CONTENT_ROOT_PATH, PACKS_FOLDER)  # full path to Packs folder in content repo
