--- Tests/scripts/collect_tests_and_content_packs.py
@@ -1,1581 +1,0 @@
-#!/usr/bin/env python3
-"""
-This script is used to create a filter_file.txt file which will run only the needed the tests for a given change.
-Overview can be found at: https://confluence.paloaltonetworks.com/display/DemistoContent/Configure+Test+Filter
-"""
-import argparse
-import glob
-import json
-import re
-from copy import deepcopy
-from packaging.version import Version
-from typing import Dict, Tuple, Optional
-
-import os
-import sys
-
-import demisto_sdk.commands.common.tools as tools
-
-from Tests.Marketplace.marketplace_services import get_last_commit_from_index
-from Tests.scripts.utils import collect_helpers
-from Tests.scripts.utils.collect_helpers import LANDING_PAGE_SECTIONS_JSON_PATH
-from Tests.scripts.utils.content_packs_util import should_test_content_pack, should_install_content_pack, \
-    is_pack_xsoar_supported
-from Tests.scripts.utils.get_modified_files_for_testing import get_modified_files_for_testing, \
-    filter_modified_files_for_specific_marketplace_version
-from Tests.scripts.utils.log_util import install_logging
-from Tests.scripts.utils import logging_wrapper as logging
-from demisto_sdk.commands.common import constants
-
-SANITY_TESTS = {
-    'Sanity Test - Playbook with integration',
-    'Sanity Test - Playbook with no integration',
-    'Sanity Test - Playbook with mocked integration',
-    'Sanity Test - Playbook with Unmockable Whois integration',
-}
-
-
-class TestConf(object):
-    __test__ = False  # required because otherwise pytest will try to run it as it has Test prefix
-
-    def __init__(self, conf: dict) -> None:
-
-        self._conf = conf
-
-    def get_skipped_integrations(self):
-        return list(self._conf['skipped_integrations'].keys())
-
-    def get_skipped_tests(self):
-        return list(self._conf['skipped_tests'].keys())
-
-    def get_private_tests(self):
-        return self._conf['private_tests']
-
-    def get_tests(self):
-        return self._conf.get('tests', {})
-
-    def get_marketplacev2_tests(self):
-        return self._conf['test_marketplacev2']
-
-    def get_test_playbook_ids(self):
-        conf_tests = self._conf['tests']
-        test_ids = []
-
-        for t in conf_tests:
-            playbook_id = t['playbookID']
-            test_ids.append(playbook_id)
-
-        return test_ids
-
-    def get_all_tested_integrations(self):
-        all_integrations = []
-        conf_tests = self._conf['tests']
-
-        for t in conf_tests:
-            if 'integrations' in t:
-                if isinstance(t['integrations'], list):
-                    all_integrations.extend(t['integrations'])
-                else:
-                    all_integrations.append(t['integrations'])
-
-        return all_integrations
-
-    def get_tested_integrations_for_collected_tests(self, collected_tests):
-        tested_integrations = []
-        conf_tests = self._conf['tests']
-
-        for t in conf_tests:
-            if t.get('playbookID') not in collected_tests:
-                continue
-
-            if 'integrations' in t:
-                if isinstance(t['integrations'], list):
-                    tested_integrations.extend(t['integrations'])
-                else:
-                    tested_integrations.append(t['integrations'])
-
-        return tested_integrations
-
-    # This function is the same function exactly as 'get_content_pack_name_of_test' and therefore should be removed
-    def get_packs_of_collected_tests(self, collected_tests, id_set):
-        packs = set([])
-        if collected_tests:
-            for test_obj in id_set.get('TestPlaybooks', []):
-                for test_id, test_data in test_obj.items():
-                    test_obj_name = test_obj[test_id].get('name')
-                    test_obj_pack = test_obj[test_id].get('pack')
-                    if test_obj_name in collected_tests and test_obj_pack:
-                        packs.add(test_obj_pack)
-        return packs
-
-    def get_packs_of_tested_integrations(self, collected_tests, id_set):
-        packs = set([])
-        tested_integrations = self.get_tested_integrations_for_collected_tests(collected_tests)
-        for integration in tested_integrations:
-            try:
-                int_path = id_set__get_integration_file_path(id_set, integration)
-                if int_path:
-                    pack = tools.get_pack_name(int_path)
-                    if pack:
-                        packs.add(pack)
-            except TypeError:
-                err_msg = f'Error occurred when trying to determine the pack of integration "{integration}"'
-                err_msg += f' with path "{int_path}"' if int_path else ''
-                logging.exception(err_msg)
-        return packs
-
-    def get_test_playbooks_configured_with_integration(self, integration_id):
-        test_playbooks = []
-        conf_tests = self._conf['tests']
-
-        for t in conf_tests:
-            if 'integrations' in t:
-                if integration_id in t['integrations']:
-                    test_playbooks.append(t['playbookID'])
-
-        return test_playbooks
-
-
-SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
-CONTENT_DIR = os.path.abspath(SCRIPT_DIR + '/../..')
-sys.path.append(CONTENT_DIR)
-
-# Global used to indicate if failed during any of the validation states
-_FAILED = False
-ID_SET = {}
-CONF: TestConf = None  # type: ignore[assignment]
-
-ARTIFACTS_FOLDER = os.getenv('ARTIFACTS_FOLDER', './artifacts')
-ARTIFACTS_ID_SET_PATH = os.path.join(ARTIFACTS_FOLDER, 'id_set.json')
-ARTIFACTS_CONF_PATH = os.path.join(ARTIFACTS_FOLDER, 'conf.json')
-if os.path.isfile(ARTIFACTS_ID_SET_PATH):
-    with open(ARTIFACTS_ID_SET_PATH, 'r') as conf_file:
-        ID_SET = json.load(conf_file)
-
-if os.path.isfile(ARTIFACTS_CONF_PATH):
-    with open(ARTIFACTS_CONF_PATH, 'r') as conf_file:
-        CONF = TestConf(json.load(conf_file))
-
-
-def is_runnable_in_server_version(from_v, server_v, to_v):
-    """
-    Checks whether an obj is runnable in a version
-    Args:
-        from_v (string): string representing Demisto version (fromversion comparable)
-        server_v (string): string representing Demisto version (version to be ran on)
-        to_v (string): string representing Demisto version (toversion comparable)
-
-    Returns:
-        bool. true if obj is runnable
-    """
-    return tools.server_version_compare(from_v, server_v) <= 0 and tools.server_version_compare(server_v, to_v) <= 0
-
-
-def get_name(file_path):
-    data_dictionary = tools.get_yaml(file_path)
-
-    if data_dictionary:
-        return data_dictionary.get('name', '-')
-
-    return None
-
-
-def get_tests(file_path):
-    """Collect tests mentioned in file_path"""
-    data_dictionary = tools.get_yaml(file_path)
-    # inject no tests to whitelist so adding values to white list will not force all tests
-    if data_dictionary:
-        return data_dictionary.get('tests', [])
-    return []
-
-
-def collect_tests_and_content_packs(
-        script_ids,
-        playbook_ids,
-        integration_ids,
-        catched_scripts,
-        catched_playbooks,
-        tests_set,
-        id_set=deepcopy(ID_SET),
-        conf=deepcopy(CONF)
-):
-    """Collect tests for the affected script_ids,playbook_ids,integration_ids.
-
-    :param script_ids: The ids of the affected scripts in your change set.
-    :param playbook_ids: The ids of the affected playbooks in your change set.
-    :param integration_ids: The ids of the affected integrations in your change set.
-    :param catched_scripts: The names of the scripts we already identified a test for.
-    :param catched_playbooks: The names of the scripts we already v a test for.
-    :param tests_set: The names of the tests we alredy identified.
-    :param id_set: The id_set json.
-    :param conf: The conf json.
-
-    :return: (test_ids, missing_ids) - All the names of possible tests, the ids we didn't match a test for.
-    """
-    caught_missing_test = False
-    catched_intergrations = set([])
-
-    test_ids = conf.get_test_playbook_ids()
-    skipped_tests = conf.get_skipped_tests()
-    skipped_integrations = conf.get_skipped_integrations()
-
-    integration_set = id_set['integrations']
-    test_playbooks_set = id_set['TestPlaybooks']
-    integration_to_command, _ = get_integration_commands(integration_ids, integration_set)
-
-    for test_playbook in test_playbooks_set:
-        detected_usage = False
-        test_playbook_id = list(test_playbook.keys())[0]
-        test_playbook_data = list(test_playbook.values())[0]
-        test_playbook_name = test_playbook_data.get('name')
-
-        for script in test_playbook_data.get('implementing_scripts', []):
-            if script in script_ids:
-                detected_usage = True
-                tests_set.add(test_playbook_id)
-                catched_scripts.add(script)
-
-        for playbook in test_playbook_data.get('implementing_playbooks', []):
-            if playbook in playbook_ids:
-                detected_usage = True
-                tests_set.add(test_playbook_id)
-                catched_playbooks.add(playbook)
-
-        if integration_to_command:
-            command_to_integration = test_playbook_data.get('command_to_integration', {})
-            for command in test_playbook_data.get('command_to_integration', {}).keys():
-                for integration_id, integration_commands in integration_to_command.items():
-                    if command in integration_commands:
-                        if not command_to_integration.get(command) or \
-                                command_to_integration.get(command) == integration_id:
-                            detected_usage = True
-                            tests_set.add(test_playbook_id)
-                            catched_intergrations.add(integration_id)
-
-        if detected_usage and test_playbook_id not in test_ids and test_playbook_id not in skipped_tests:
-            caught_missing_test = check_if_test_should_not_be_missed(test_playbook_data.get('file_path', ''),
-                                                                     test_playbook_name)
-
-    ids_with_no_tests = update_missing_sets(catched_intergrations, catched_playbooks, catched_scripts,
-                                            integration_ids, playbook_ids, script_ids)
-
-    # remove skipped integrations from the list
-    ids_with_no_tests = ids_with_no_tests - set(skipped_integrations)
-    packs_to_install = set()
-    id_set_test_playbooks = id_set.get('TestPlaybooks', [])
-    for test_playbook in id_set_test_playbooks:
-        test_playbook_id = list(test_playbook.keys())[0]
-        test_playbook_object = test_playbook[test_playbook_id]
-        if test_playbook_id in tests_set:
-            test_playbook_pack = test_playbook_object.get('pack')
-            if test_playbook_pack:
-                logging.info(
-                    f'Found test playbook "{test_playbook_id}" in pack "{test_playbook_pack}"'
-                    f' - adding to packs to install')
-                packs_to_install.add(test_playbook_pack)
-            else:
-                logging.warning(f'Found test playbook "{test_playbook_id}" without pack'
-                                f' - not adding to packs to install')
-
-    return test_ids, ids_with_no_tests, caught_missing_test, packs_to_install
-
-
-def check_if_test_should_not_be_missed(test_playbook_path: str, test_playbook_name: str) -> bool:
-    """
-    Checks if a missing test from conf JSON should be missed or not.
-    Test should be missed if the support level is not XSOAR.
-    Args:
-        test_playbook_path (str): Path to test playbook.
-        test_playbook_name (str): Name of the test playbook. For logging purposes.
-
-    Returns:
-        (bool): Whether test should be missed.
-    """
-    pack_name: Optional[str] = tools.get_pack_name(test_playbook_path)
-    if is_pack_xsoar_supported(test_playbook_path):
-        logging.error(f'The playbook "{test_playbook_name}" does not appear in the conf.json file,'
-                      " which means no test with it will run. please update the conf.json file accordingly."
-                      )
-        return True
-    else:
-        logging.info(f'The playbook "{test_playbook_name}" does not appear in conf.json file, but pack '
-                     f'{pack_name} is not XSOAR supported. Not failing test collection.')
-        return False
-
-
-def update_missing_sets(catched_intergrations, catched_playbooks, catched_scripts, integration_ids, playbook_ids,
-                        script_ids):
-    missing_integrations = integration_ids - catched_intergrations
-    missing_playbooks = playbook_ids - catched_playbooks
-    missing_scripts = script_ids - catched_scripts
-    missing_ids = missing_integrations.union(missing_playbooks).union(missing_scripts)
-    return missing_ids
-
-
-def get_integration_commands(integration_ids, integration_set):
-    integration_to_command = {}
-    deprecated_message = ''
-    deprecated_commands_string = ''
-    for integration in integration_set:
-        integration_id = list(integration.keys())[0]
-        integration_data = list(integration.values())[0]
-        if integration_id in integration_ids:
-            integration_commands = set(integration_data.get('commands', []))
-            integration_deprecated_commands = set(integration_data.get('deprecated_commands', []))
-            if integration_deprecated_commands:
-                deprecated_names = ', '.join(integration_deprecated_commands)
-                deprecated_commands_string += '{}: {}\n'.format(integration_id, deprecated_names)
-
-            relevant_commands = list(integration_commands - integration_deprecated_commands)
-            integration_to_command[integration_id] = relevant_commands
-
-    if deprecated_commands_string:
-        deprecated_message = 'The following integration commands are deprecated and are not taken ' \
-                             'into account in the test collection:\n{}'.format(deprecated_commands_string)
-
-    return integration_to_command, deprecated_message
-
-
-def is_integration_fetching_incidents(integration_yml_path):
-    integration_yml_dict = tools.get_yaml(integration_yml_path)
-
-    return integration_yml_dict.get('script').get('isfetch', False) is True
-
-
-def id_set__get_test_playbook(id_set, test_playbook_id):
-    for test_playbook in id_set.get('TestPlaybooks', []):
-        if test_playbook_id in test_playbook.keys():
-            return test_playbook[test_playbook_id]
-
-    return None
-
-
-def id_set__get_integration_file_path(id_set, integration_id):
-    for integration in id_set.get('integrations', []):
-        if integration_id in integration.keys():
-            return integration[integration_id]['file_path']
-
-    logging.critical(f'Could not find integration "{integration_id}" in the id_set')
-    return None
-
-
-def check_if_fetch_incidents_is_tested(missing_ids, integration_ids, id_set, conf, tests_set):
-    # If integration is mentioned/used in one of the test configurations, it means that integration is tested.
-    # For example there could be a test playbook that tests fetch incidents command of some integration
-    # so the test playbook will use FetchFromInstance script in the playbook, which is not direct command of a specific
-    # integration
-
-    missing_integration_ids = missing_ids & integration_ids
-    for missing_id in missing_integration_ids:
-        integration_file_path = id_set__get_integration_file_path(id_set, missing_id)
-        is_fetching = is_integration_fetching_incidents(integration_file_path)
-        if not is_fetching:
-            continue
-
-        test_playbook_ids = conf.get_test_playbooks_configured_with_integration(missing_id)
-        if len(test_playbook_ids) == 0:
-            # there are no test playbooks for this integration configured
-            continue
-
-        for test_playbook_id in test_playbook_ids:
-            test_playbook = id_set__get_test_playbook(id_set, test_playbook_id)
-            if test_playbook and 'FetchFromInstance' in test_playbook.get('implementing_scripts', []):
-                missing_ids = missing_ids - {missing_id}
-                tests_set.add(test_playbook_id)
-
-    return missing_ids, tests_set
-
-
-def find_tests_and_content_packs_for_modified_files(modified_files, conf=deepcopy(CONF), id_set=deepcopy(ID_SET)):
-    script_names: set = set([])
-    playbook_names: set = set([])
-    integration_ids: set = set([])
-
-    tests_set, caught_scripts, caught_playbooks, packs_to_install = collect_changed_ids(
-        integration_ids, playbook_names, script_names, modified_files, id_set)
-
-    test_ids, missing_ids, caught_missing_test, test_packs_to_install = collect_tests_and_content_packs(
-        script_names, playbook_names, integration_ids, caught_scripts, caught_playbooks, tests_set, id_set, conf)
-
-    packs_to_install.update(test_packs_to_install)
-
-    missing_ids = update_with_tests_sections(missing_ids, modified_files, test_ids, tests_set)
-
-    missing_ids, tests_set = check_if_fetch_incidents_is_tested(missing_ids, integration_ids, id_set, conf, tests_set)
-
-    if len(missing_ids) > 0:
-        test_string = '\n'.join(missing_ids)
-        message = "Was not able to find tests for:\n{0}".format(test_string)
-        logging.error(message)
-
-    if caught_missing_test or len(missing_ids) > 0:
-        global _FAILED
-        _FAILED = True
-
-    return tests_set, packs_to_install
-
-
-def update_with_tests_sections(missing_ids, modified_files, test_ids, tests):
-    # Search for tests section
-    for file_path in modified_files:
-        tests_from_file = get_tests(file_path)
-        for test in tests_from_file:
-            if collect_helpers.checked_type(file_path, collect_helpers.INTEGRATION_REGEXES):
-                _id = tools.get_script_or_integration_id(file_path)
-
-            else:
-                _id = get_name(file_path)
-            if test in test_ids or re.match(collect_helpers.NO_TESTS_FORMAT, test, re.IGNORECASE):
-
-                missing_ids = missing_ids - {_id}
-                tests.add(test)
-
-            else:
-                if check_if_test_should_not_be_missed(file_path, test):
-                    global _FAILED
-                    _FAILED = True
-                else:
-                    missing_ids = missing_ids - {_id}
-
-    return missing_ids
-
-
-def collect_content_packs_to_install(id_set: Dict, integration_ids: set, playbook_names: set, script_names: set) -> set:
-    """Iterates all content entities in the ID set and extract the pack names for the modified ones.
-
-    Args:
-        id_set (Dict): Structure which holds all content entities to extract pack names from.
-        integration_ids (set): Set of integration IDs to get pack names for.
-        playbook_names (set): Set of playbook names to get pack names for.
-        script_names (set): Set of script names to get pack names for.
-
-    Returns:
-        set. Pack names to install.
-    """
-    packs_to_install = set()
-
-    id_set_integrations = id_set.get('integrations', [])
-    for integration in id_set_integrations:
-        integration_id = list(integration.keys())[0]
-        integration_object = integration[integration_id]
-        if integration_id in integration_ids:
-            integration_pack = integration_object.get('pack')
-            if integration_pack:
-                logging.info(
-                    f'Found integration {integration_id} in pack {integration_pack} - adding to packs to install')
-                packs_to_install.add(integration_object.get('pack'))
-            else:
-                logging.warning(f'Found integration {integration_id} without pack - not adding to packs to install')
-
-    id_set_playbooks = id_set.get('playbooks', [])
-    for playbook in id_set_playbooks:
-        playbook_object = list(playbook.values())[0]
-        playbook_name = playbook_object.get('name')
-        if playbook_name in playbook_names:
-            playbook_pack = playbook_object.get('pack')
-            if playbook_pack:
-                logging.info(f'Found playbook {playbook_name} in pack {playbook_pack} - adding to packs to install')
-                packs_to_install.add(playbook_pack)
-            else:
-                logging.warning(f'Found playbook {playbook_name} without pack - not adding to packs to install')
-
-    id_set_script = id_set.get('scripts', [])
-    for script in id_set_script:
-        script_id = list(script.keys())[0]
-        script_object = script[script_id]
-        if script_id in script_names:
-            script_pack = script_object.get('pack')
-            if script_pack:
-                logging.info(f'Found script {script_id} in pack {script_pack} - adding to packs to install')
-                packs_to_install.add(script_object.get('pack'))
-            else:
-                logging.warning(f'Found script {script_id} without pack - not adding to packs to install')
-
-    return packs_to_install
-
-
-def get_api_module_integrations(changed_api_modules, integration_set):
-    integration_to_version = {}
-    integration_ids_to_test = set([])
-    for integration in integration_set:
-        integration_data = list(integration.values())[0]
-        if changed_api_modules & set(integration_data.get('api_modules', [])):
-            file_path = integration_data.get('file_path')
-            integration_id = tools.get_script_or_integration_id(file_path)
-            integration_ids_to_test.add(integration_id)
-            integration_to_version[integration_id] = (tools.get_from_version(file_path),
-                                                      tools.get_to_version(file_path))
-
-    return integration_ids_to_test, integration_to_version
-
-
-def collect_changed_ids(integration_ids, playbook_names, script_names, modified_files, id_set=deepcopy(ID_SET)):
-    tests_set: set = set([])
-    updated_script_names: set = set([])
-    updated_playbook_names: set = set([])
-    catched_scripts: set = set([])
-    catched_playbooks: set = set([])
-    changed_api_modules: set = set([])
-
-    script_to_version = {}
-    playbook_to_version = {}
-    integration_to_version = {}
-    for file_path in modified_files:
-        if collect_helpers.checked_type(file_path, collect_helpers.SCRIPT_REGEXES + constants.YML_SCRIPT_REGEXES):
-            name = get_name(file_path)
-            script_names.add(name)
-            script_to_version[name] = (tools.get_from_version(file_path), tools.get_to_version(file_path))
-
-            package_name = os.path.dirname(file_path)
-            if glob.glob(package_name + "/*_test.py"):
-                catched_scripts.add(name)
-                tests_set.add('Found a unittest for the script {}'.format(package_name))
-
-        elif collect_helpers.checked_type(file_path, constants.YML_PLAYBOOKS_NO_TESTS_REGEXES):
-            name = get_name(file_path)
-            playbook_names.add(name)
-            playbook_to_version[name] = (tools.get_from_version(file_path), tools.get_to_version(file_path))
-
-        elif collect_helpers.checked_type(file_path, collect_helpers.INTEGRATION_REGEXES + constants.YML_INTEGRATION_REGEXES):
-            _id = tools.get_script_or_integration_id(file_path)
-            integration_ids.add(_id)
-            integration_to_version[_id] = (tools.get_from_version(file_path), tools.get_to_version(file_path))
-
-        if collect_helpers.checked_type(file_path, constants.API_MODULE_REGEXES):
-            api_module_name = tools.get_script_or_integration_id(file_path)
-            changed_api_modules.add(api_module_name)
-
-    script_set = id_set['scripts']
-    playbook_set = id_set['playbooks']
-    integration_set = id_set['integrations']
-
-    if changed_api_modules:
-        integration_ids_to_test, integration_to_version_to_add = get_api_module_integrations(changed_api_modules,
-                                                                                             integration_set)
-        integration_ids = integration_ids.union(integration_ids_to_test)
-        integration_to_version = {**integration_to_version, **integration_to_version_to_add}
-
-    deprecated_msgs = exclude_deprecated_entities(script_set, script_names,
-                                                  playbook_set, playbook_names,
-                                                  integration_set, integration_ids)
-
-    for script_id in script_names:
-        enrich_for_script_id(script_id, script_to_version[script_id], script_names, script_set, playbook_set,
-                             playbook_names, updated_script_names, updated_playbook_names, catched_scripts,
-                             catched_playbooks, tests_set)
-
-    integration_to_command, deprecated_commands_message = get_integration_commands(integration_ids, integration_set)
-    for integration_id, integration_commands in integration_to_command.items():
-        enrich_for_integration_id(integration_id, integration_to_version[integration_id], integration_commands,
-                                  script_set, playbook_set, playbook_names, script_names, updated_script_names,
-                                  updated_playbook_names, catched_scripts, catched_playbooks, tests_set)
-
-    for playbook_id in playbook_names:
-        enrich_for_playbook_id(playbook_id, playbook_to_version[playbook_id], playbook_names, script_set, playbook_set,
-                               updated_playbook_names, catched_playbooks, tests_set)
-
-    for new_script in updated_script_names:
-        script_names.add(new_script)
-
-    for new_playbook in updated_playbook_names:
-        playbook_names.add(new_playbook)
-
-    affected_ids_strings = {
-        'scripts': '',
-        'playbooks': '',
-        'integrations': ''
-    }
-    if script_names:
-        affected_ids_strings['scripts'] += 'Scripts:\n' + '\n'.join(script_names)
-    if playbook_names:
-        affected_ids_strings['playbooks'] += 'Playbooks:\n' + '\n'.join(playbook_names)
-    if integration_ids:
-        affected_ids_strings['integrations'] += 'Integrations:\n' + '\n'.join(integration_ids)
-
-    logging.info('The following ids are affected due to the changes you made:')
-    for entity in ['scripts', 'playbooks', 'integrations']:
-        logging.info(affected_ids_strings[entity])
-        logging.warning(deprecated_msgs[entity])
-
-    if deprecated_commands_message:
-        logging.warning(deprecated_commands_message)
-
-    packs_to_install = collect_content_packs_to_install(id_set, integration_ids, playbook_names, script_names)
-    return tests_set, catched_scripts, catched_playbooks, packs_to_install
-
-
-def exclude_deprecated_entities(script_set, script_names,
-                                playbook_set, playbook_names,
-                                integration_set, integration_ids):
-    """Removes deprecated entities from the affected entities sets.
-
-    :param script_set: The set of existing scripts within Content repo.
-    :param script_names: The names of the affected scripts in your change set.
-    :param playbook_set: The set of existing playbooks within Content repo.
-    :param playbook_names: The ids of the affected playbooks in your change set.
-    :param integration_set: The set of existing integrations within Content repo.
-    :param integration_ids: The ids of the affected integrations in your change set.
-
-    :return: deprecated_messages_dict - A dict of messages specifying of all the deprecated entities.
-    """
-    deprecated_messages_dict = {
-        'scripts': '',
-        'playbooks': '',
-        'integrations': ''
-    }
-
-    deprecated_entities_strings_dict = {
-        'scripts': '',
-        'playbooks': '',
-        'integrations': ''
-    }
-
-    # Iterates over three types of entities: scripts, playbooks and integrations and removes deprecated entities
-    for entity_set, entity_names, entity_type in [(script_set, script_names, 'scripts'),
-                                                  (playbook_set, playbook_names, 'playbooks'),
-                                                  (integration_set, integration_ids, 'integrations')]:
-        for entity in entity_set:
-            # integrations are defined by their ids while playbooks and scripts and scripts are defined by names
-            if entity_type == 'integrations':
-                entity_name = list(entity.keys())[0]
-            else:
-                entity_name = list(entity.values())[0].get('name', '')
-
-            if entity_name in entity_names:
-                entity_data = list(entity.values())[0]
-                if entity_data.get('deprecated', False):
-                    deprecated_entities_strings_dict[entity_type] += entity_name + '\n'
-                    entity_names.remove(entity_name)
-
-        if deprecated_entities_strings_dict[entity_type]:
-            deprecated_messages_dict[entity_type] = 'The following {} are deprecated ' \
-                                                    'and are not taken into account in the test collection:' \
-                                                    '\n{}'.format(entity_type,
-                                                                  deprecated_entities_strings_dict[entity_type])
-
-    return deprecated_messages_dict
-
-
-def enrich_for_integration_id(integration_id, given_version, integration_commands, script_set, playbook_set,
-                              playbook_names, script_names, updated_script_names, updated_playbook_names,
-                              catched_scripts, catched_playbooks, tests_set):
-    """Enrich the list of affected scripts/playbooks by your change set.
-
-    :param integration_id: The name of the integration we changed.
-    :param given_version: the version of the integration we changed.
-    :param integration_commands: The commands of the changed integation
-    :param script_set: The set of existing scripts within Content repo.
-    :param playbook_set: The set of existing playbooks within Content repo.
-    :param playbook_names: The names of the playbooks affected by your changes.
-    :param script_names: The names of the scripts affected by your changes.
-    :param updated_script_names: The names of scripts we identify as affected to your change set.
-    :param updated_playbook_names: The names of playbooks we identify as affected to your change set.
-    :param catched_scripts: The names of scripts we found tests for.
-    :param catched_playbooks: The names of playbooks we found tests for.
-    :param tests_set: The names of the caught tests.
-    """
-    for playbook in playbook_set:
-        playbook_data = list(playbook.values())[0]
-        if playbook_data.get('deprecated', False):
-            continue
-        playbook_name = playbook_data.get('name')
-        playbook_fromversion = playbook_data.get('fromversion', '0.0.0')
-        playbook_toversion = playbook_data.get('toversion', '99.99.99')
-        command_to_integration = playbook_data.get('command_to_integration', {})
-        implementing_commands = command_to_integration.keys()
-        for integration_command in integration_commands:
-            if integration_command in implementing_commands and playbook_toversion >= given_version[1]:
-                if playbook_name not in playbook_names and playbook_name not in updated_playbook_names:
-                    if not command_to_integration.get(integration_command) or \
-                            command_to_integration.get(integration_command) == integration_id:
-
-                        tests = set(playbook_data.get('tests', []))
-                        if tests:
-                            catched_playbooks.add(playbook_name)
-                            update_test_set(tests, tests_set)
-
-                        updated_playbook_names.add(playbook_name)
-                        new_versions = (playbook_fromversion, playbook_toversion)
-                        enrich_for_playbook_id(playbook_name, new_versions, playbook_names, script_set, playbook_set,
-                                               updated_playbook_names, catched_playbooks, tests_set)
-
-    for script in script_set:
-        script_data = list(script.values())[0]
-        if script_data.get('deprecated', False):
-            continue
-        script_name = script_data.get('name')
-        script_file_path = script_data.get('file_path')
-        script_fromversion = script_data.get('fromversion', '0.0.0')
-        script_toversion = script_data.get('toversion', '99.99.99')
-        command_to_integration = script_data.get('command_to_integration', {})
-        for integration_command in integration_commands:
-            if integration_command in script_data.get('depends_on', []):
-                if integration_command in command_to_integration.keys() and \
-                        command_to_integration[integration_command] == integration_id and \
-                        script_toversion >= given_version[1]:
-
-                    if script_name not in script_names and script_name not in updated_script_names:
-                        tests = script_data.get('tests', [])
-                        if tests:
-                            catched_scripts.add(script_name)
-                            update_test_set(tests, tests_set)
-
-                        package_name = os.path.dirname(script_file_path)
-                        if glob.glob(package_name + "/*_test.py"):
-                            catched_scripts.add(script_name)
-                            tests_set.add('Found a unittest for the script {}'.format(script_name))
-
-                        updated_script_names.add(script_name)
-                        new_versions = (script_fromversion, script_toversion)
-                        enrich_for_script_id(script_name, new_versions, script_names, script_set, playbook_set,
-                                             playbook_names, updated_script_names, updated_playbook_names,
-                                             catched_scripts, catched_playbooks, tests_set)
-
-
-def enrich_for_playbook_id(given_playbook_id, given_version, playbook_names, script_set, playbook_set,
-                           updated_playbook_names, catched_playbooks, tests_set):
-    for playbook in playbook_set:
-        playbook_data = list(playbook.values())[0]
-        if playbook_data.get('deprecated', False):
-            continue
-        playbook_name = playbook_data.get('name')
-        playbook_fromversion = playbook_data.get('fromversion', '0.0.0')
-        playbook_toversion = playbook_data.get('toversion', '99.99.99')
-        if given_playbook_id in playbook_data.get('implementing_playbooks', []) and \
-                playbook_toversion >= given_version[1]:
-
-            if playbook_name not in playbook_names and playbook_name not in updated_playbook_names:
-                tests = set(playbook_data.get('tests', []))
-                if tests:
-                    catched_playbooks.add(playbook_name)
-                    update_test_set(tests, tests_set)
-
-                updated_playbook_names.add(playbook_name)
-                new_versions = (playbook_fromversion, playbook_toversion)
-                enrich_for_playbook_id(playbook_name, new_versions, playbook_names, script_set, playbook_set,
-                                       updated_playbook_names, catched_playbooks, tests_set)
-
-
-def enrich_for_script_id(given_script_id, given_version, script_names, script_set, playbook_set, playbook_names,
-                         updated_script_names, updated_playbook_names, catched_scripts, catched_playbooks, tests_set):
-    for script in script_set:
-        script_data = list(script.values())[0]
-        if script_data.get('deprecated', False):
-            continue
-        script_name = script_data.get('name')
-        script_file_path = script_data.get('file_path')
-        script_fromversion = script_data.get('fromversion', '0.0.0')
-        script_toversion = script_data.get('toversion', '99.99.99')
-        if given_script_id in script_data.get('script_executions', []) and script_toversion >= given_version[1]:
-            if script_name not in script_names and script_name not in updated_script_names:
-                tests = set(script_data.get('tests', []))
-                if tests:
-                    catched_scripts.add(script_name)
-                    update_test_set(tests, tests_set)
-
-                package_name = os.path.dirname(script_file_path)
-                if glob.glob(package_name + "/*_test.py"):
-                    catched_scripts.add(script_name)
-                    tests_set.add('Found a unittest for the script {}'.format(script_name))
-
-                updated_script_names.add(script_name)
-                new_versions = (script_fromversion, script_toversion)
-                enrich_for_script_id(script_name, new_versions, script_names, script_set, playbook_set, playbook_names,
-                                     updated_script_names, updated_playbook_names, catched_scripts, catched_playbooks,
-                                     tests_set)
-
-    for playbook in playbook_set:
-        playbook_data = list(playbook.values())[0]
-        if playbook_data.get('deprecated', False):
-            continue
-        playbook_name = playbook_data.get('name')
-        playbook_fromversion = playbook_data.get('fromversion', '0.0.0')
-        playbook_toversion = playbook_data.get('toversion', '99.99.99')
-        if given_script_id in playbook_data.get('implementing_scripts', []) and playbook_toversion >= given_version[1]:
-            if playbook_name not in playbook_names and playbook_name not in updated_playbook_names:
-                tests = set(playbook_data.get('tests', []))
-                if tests:
-                    catched_playbooks.add(playbook_name)
-                    update_test_set(tests, tests_set)
-
-                updated_playbook_names.add(playbook_name)
-                new_versions = (playbook_fromversion, playbook_toversion)
-                enrich_for_playbook_id(playbook_name, new_versions, playbook_names, script_set, playbook_set,
-                                       updated_playbook_names, catched_playbooks, tests_set)
-
-
-def update_test_set(tests, tests_set):
-    for test in tests:
-        tests_set.add(test)
-
-
-def get_test_conf_from_conf(test_id, server_version, conf=deepcopy(CONF)):
-    """Gets first occurrence of test conf with matching playbookID value to test_id with a valid from/to version"""
-    test_conf_lst = conf.get_tests()
-    # return None if nothing is found
-    test_conf = next((test_conf for test_conf in test_conf_lst if
-                      (test_conf.get('playbookID') == test_id and is_runnable_in_server_version(
-                          from_v=test_conf.get('fromversion', '0.0'),
-                          server_v=server_version,
-                          to_v=test_conf.get('toversion', '99.99.99')))), None)
-    return test_conf
-
-
-def extract_matching_object_from_id_set(obj_id, obj_set, server_version='0'):
-    """Gets first occurrence of object in the object's id_set with matching id/name and valid from/to version"""
-    for obj_wrpr in obj_set:
-        # try to get object by id
-        if obj_id in obj_wrpr:
-            obj = obj_wrpr.get(obj_id)
-
-        # try to get object by name
-        else:
-            obj_keys = list(obj_wrpr.keys())
-            if not obj_keys:
-                continue
-            obj = obj_wrpr[obj_keys[0]]
-            if obj.get('name') != obj_id:
-                continue
-
-        # check if object is runnable
-        fromversion = obj.get('fromversion', '0.0')
-        toversion = obj.get('toversion', '99.99.99')
-        if is_runnable_in_server_version(from_v=fromversion, server_v=server_version, to_v=toversion):
-            return obj
-    return None
-
-
-def get_packs_from_landing_page(branch_name: str) -> set:
-    """
-    Parses the changes made by the current branch to the landing page sections file into modified packs.
-    These modified packs will have the metadata files changed.
-    Args:
-        branch_name: The name of the branch to compare changes from origin/master
-
-    Returns:
-        A set of packs that needs to be updated.
-    """
-    with open(LANDING_PAGE_SECTIONS_JSON_PATH, 'r') as file:
-        landing_page_sections_content = json.load(file)
-    landing_page_sections_keys = landing_page_sections_content.keys()
-    change_string = tools.run_command(f'git diff origin/master...{branch_name} {LANDING_PAGE_SECTIONS_JSON_PATH}')
-    changed_packs = set()
-    for changed_group in re.findall(r'[-+][ ]+"(.*)"', change_string):
-        if changed_group not in landing_page_sections_keys:
-            changed_packs.add(changed_group)
-    return changed_packs
-
-
-def get_test_from_conf(branch_name, conf=deepcopy(CONF)):
-    tests = set([])
-    changed = set([])
-    change_string = tools.run_command("git diff origin/master...{} Tests/conf.json".format(branch_name))
-    added_groups = re.findall(r'(\+[ ]+")(.*)(":)', change_string)
-    if added_groups:
-        for group in added_groups:
-            changed.add(group[1])
-
-    deleted_groups = re.findall(r'(-[ ]+")(.*)(":)', change_string)
-    if deleted_groups:
-        for group in deleted_groups:
-            changed.add(group[1])
-
-    conf_tests = conf.get_tests()
-    for t in conf_tests:
-        playbook_id = t['playbookID']
-        integrations_conf = t.get('integrations', [])
-        if playbook_id in changed:
-            tests.add(playbook_id)
-            continue
-
-        if not isinstance(integrations_conf, list):
-            integrations_conf = [integrations_conf]
-
-        for integration in integrations_conf:
-            if integration in changed:
-                tests.add(playbook_id)
-
-    if not tests:
-        tests.add('changed skip section')
-
-    return tests
-
-
-def is_test_runnable(test_id, id_set, conf, server_version):
-    """Checks whether the test is runnable
-    1. Test is not skipped.
-    2. Test playbook / integration is not skipped.
-    3. Test fromversion is earlier or equal to server_version
-    4. Test toversion is greater or equal to server_version
-    5. If test has integrations, then all integrations
-        a. fromversion is earlier or equal to server_version
-        b. toversion is after or equal to server_version
-    """
-    skipped_tests = conf.get_skipped_tests()
-    warning_prefix = f'{test_id} is not runnable on {server_version}'
-    # check if test is skipped
-    if test_id in skipped_tests:
-        logging.debug(f'{warning_prefix} - skipped')
-        return False
-    test_conf = get_test_conf_from_conf(test_id, server_version, conf)
-
-    # check if there's a test to run
-    if not test_conf:
-        logging.debug(f'{warning_prefix} - couldn\'t find test in conf.json')
-        return False
-    conf_fromversion = test_conf.get('fromversion', '0.0')
-    conf_toversion = test_conf.get('toversion', '99.99.99')
-    test_playbooks_set = id_set.get('TestPlaybooks', [])
-    test_playbook_obj = extract_matching_object_from_id_set(test_id, test_playbooks_set, server_version)
-
-    # check whether the test is runnable in id_set
-    if not test_playbook_obj:
-        logging.debug(f'{warning_prefix} - couldn\'t find the test in id_set.json')
-        return False
-
-    # check used integrations available
-    if not is_test_integrations_available(server_version, test_conf, conf, id_set):
-        logging.debug(f'{warning_prefix} - no active integration found')
-        return False
-
-    # check conf from/to
-    if not is_runnable_in_server_version(conf_fromversion, server_version, conf_toversion):
-        logging.debug(f'{warning_prefix} - conf.json from/to version')
-        return False
-
-    return True
-
-
-def is_test_integrations_available(server_version, test_conf, conf, id_set):
-    """
-    Check if all used integrations are skipped / available
-    """
-    test_integration_ids = test_conf.get('integrations')
-    if test_integration_ids:
-        if not isinstance(test_integration_ids, list):
-            test_integration_ids = [test_integration_ids]
-        if not is_test_uses_active_integration(test_integration_ids, conf):
-            return False
-        # check if all integration from/toversion is valid with server_version
-        integrations_set = id_set.get('integrations', [])
-        if any(extract_matching_object_from_id_set(integration_id, integrations_set, server_version) is None for
-               integration_id in
-               test_integration_ids):
-            return False
-    return True
-
-
-def is_test_uses_active_integration(integration_ids, conf=deepcopy(CONF)):
-    """Checks whether there's an an integration in test_integration_ids that's not skipped"""
-    skipped_integrations = conf.get_skipped_integrations()
-    # check if all integrations are skipped
-    if all(integration_id in skipped_integrations for integration_id in integration_ids):
-        return False
-
-    return True
-
-
-def get_tests_for_pack(pack_path):
-    pack_yml_files = tools.get_files_in_dir(pack_path, ['yml'])
-    pack_test_playbooks = [tools.collect_ids(file) for file in pack_yml_files if
-                           collect_helpers.checked_type(file, constants.YML_TEST_PLAYBOOKS_REGEXES)]
-    return pack_test_playbooks
-
-
-def get_content_pack_name_of_test(tests: set, id_set: Optional[Dict] = None) -> set:
-    """Returns the content packs names in which given test playbooks are in.
-
-    Args:
-        tests (set): The names of the tests to find their content packs.
-        id_set (Dict): Structure which holds all content entities to extract pack names from.
-
-    Returns:
-        str. The content pack name in which the test playbook is in.
-    """
-    content_packs = set()
-    if id_set is not None:
-        for test_playbook_object in id_set.get('TestPlaybooks', []):
-            test_playbook_name = list(test_playbook_object.keys())[0]
-            test_playbook_data = list(test_playbook_object.values())[0]
-            if test_playbook_name in tests:
-                pack_name = test_playbook_data.get('pack')
-                if pack_name:
-                    content_packs.add(pack_name)
-                    if len(tests) == len(content_packs):
-                        # we found all content packs for all tests we were looking for
-                        break
-
-    return content_packs
-
-
-def get_modified_packs(files_string):
-    modified_packs = set()
-    all_files = files_string.splitlines()
-
-    for _file in all_files:
-        file_data = _file.split()
-        if not file_data:
-            continue
-        file_status = file_data[0]
-        if file_status.lower().startswith('r'):
-            file_path = file_data[2]
-        else:
-            file_path = file_data[1]
-
-        if file_path.startswith('Documentation'):
-            modified_packs.add('Base')
-
-        elif file_path.startswith('Packs'):
-            modified_packs.add(tools.get_pack_name(file_path))
-
-    return modified_packs
-
-
-def remove_ignored_tests(tests: set, id_set: dict, modified_packs: set) -> set:
-    """Filters out test playbooks, which are in .pack-ignore, from the given tests set
-    Args:
-        tests (set): Tests set to remove the tests to ignore from
-        id_set (dict): The id set object
-        modified_packs (set): List of the modified packs
-    Return:
-         set: The filtered tests set
-    """
-    ignored_tests_set = set()
-    content_packs = get_content_pack_name_of_test(tests, id_set)
-    for pack in modified_packs:
-        ignored_tests_set.update(tools.get_ignore_pack_skipped_tests(pack, modified_packs, id_set))
-    for pack in content_packs:
-        test = tools.get_ignore_pack_skipped_tests(pack, modified_packs, id_set)
-        ignored_tests_set.update(test)
-
-    if ignored_tests_set:
-        readable_ignored_tests = "\n".join(map(str, ignored_tests_set))
-        logging.info(f"Skipping tests that were ignored via .pack-ignore:\n{readable_ignored_tests}")
-        tests.difference_update(ignored_tests_set)
-
-    return tests
-
-
-def remove_tests_for_non_supported_packs(tests: set, id_set: dict, marketplace_version: str) -> set:
-    """Filters out test playbooks, which belong to one of the following packs:
-        - Non XSOAR supported packs
-        - DeprecatedContent, NonSupported packs
-        - Deprecated packs
-        - Not compatible with the marketplace version were the tests will run
-
-        Args:
-            tests (set): Tests set to remove the tests to ignore from
-            id_set (dict): The id set object
-            marketplace_version (str): The marketplace version were the tests will run
-
-        Return:
-             set: The filtered tests set
-        """
-    tests_that_should_not_be_tested = set()
-    for test in tests:
-        content_pack_name_list = list(get_content_pack_name_of_test({test}, id_set))
-        if content_pack_name_list:
-            id_set_test_playbook_pack_name = content_pack_name_list[0]
-
-            # We don't want to test playbooks that are not XSOAR supported, belong to the ignored packs or from
-            # deprecated packs.
-            should_test, reason = should_test_content_pack(id_set_test_playbook_pack_name, marketplace_version, id_set)
-            if not should_test:
-                tests_that_should_not_be_tested.add(f'{test}: {reason}')
-        else:
-            tests_that_should_not_be_tested.add(f'{test}: the pack of the test was not found in the id set')
-
-    if tests_that_should_not_be_tested:
-        logging.debug('The following test playbooks are not supported and will not be tested: \n{} '.format(
-            '\n'.join(tests_that_should_not_be_tested)))
-        tests_names = {test.split(':')[0] for test in tests_that_should_not_be_tested}
-        tests.difference_update(tests_names)
-    return tests
-
-
-def remove_private_tests(tests_without_private_packs):
-    """Remove all tests that test private integrations from the test list.
-
-    Args:
-        tests_without_private_packs: The list of tests to remove the private packs' tests from.
-    """
-    private_tests = CONF.get_private_tests()
-    for private_test in private_tests:
-        if private_test in tests_without_private_packs:
-            tests_without_private_packs.remove(private_test)
-
-
-def remove_unsupported_marketplace_tests_version(tests: set, id_set: dict, marketplace_version: str) -> set:
-    """Filters out test playbooks, which is not supported in the marketplace_version:
-        Args:
-            tests (set): Tests set to remove the tests to ignore from
-            id_set (dict): The id set object
-            marketplace_version (str): The marketplace version were the tests will run
-
-        Return:
-             set: The filtered tests set
-        """
-    ignored_tests_set = set()
-    for test_obj in id_set.get('TestPlaybooks', []):
-        for test_id, test_data in test_obj.items():
-            test_obj_name = test_data.get('name')
-            test_obj_versions = test_data.get('marketplaces', [])
-            if test_obj_name in tests and marketplace_version not in test_obj_versions:
-                ignored_tests_set.add(test_obj_name)
-
-    if ignored_tests_set:
-        readable_ignored_tests = "\n".join(map(str, ignored_tests_set))
-        logging.info(f"Skipping tests that are not supported in this marketplace_version:\n{readable_ignored_tests}")
-        tests.difference_update(ignored_tests_set)
-    return tests
-
-
-def filter_tests(tests: set, id_set: dict, modified_packs: set, marketplace_version: str, is_nightly=False) -> set:
-    """
-    Filter tests out from the test set if they are:
-    a. Ignored
-    b. Non-XSOAR or non-supported packs
-    c. tests of deprecated packs.
-    d. tests of private packs (optional)
-    e. Not compatible with the marketplace version were the tests will run
-
-    Args:
-        tests (set): Set of tests collected so far.
-        id_set (dict): The ID set.
-        modified_packs: The modified packs.
-        marketplace_version (str): The marketplace version were the tests will run.
-    Returns:
-        (set): Set of tests without ignored, non supported and deprecated-packs tests.
-    """
-    tests_with_no_dummy_strings = {test for test in tests if 'no test' not in test.lower()}
-    tests_without_ignored = remove_ignored_tests(tests_with_no_dummy_strings, id_set, modified_packs)
-    tests_without_non_supported = remove_unsupported_marketplace_tests_version(tests_without_ignored, id_set,
-                                                                               marketplace_version)
-    tests_without_non_supported_packs = remove_tests_for_non_supported_packs(tests_without_non_supported, id_set,
-                                                                             marketplace_version)
-
-    if is_nightly:
-        # Removing private packs' tests from nightly, since they aren't runnable in nightly
-        # due to the fact they aren't in stored in the content repository.
-        remove_private_tests(tests_without_non_supported_packs)
-
-    return tests_without_non_supported_packs
-
-
-def filter_installed_packs(packs_to_install: set, marketplace_version: str, id_set: dict) -> set:
-    """
-    Filter only the packs that should get installed by the following conditions:
-        - Content pack is not in skipped packs
-        - Content pack is not deprecated
-    Args:
-        packs_to_install (set): Set of installed packs collected so far.
-        marketplace_version (str): The marketplace version where the tests will run
-        id_set (dict): Structure which holds all content entities to extract pack names from.
-    Returns:
-        (set): Set of packs without ignored, skipped and deprecated-packs.
-    """
-
-    packs_that_should_not_be_installed = set()
-    packs_that_should_be_installed = set()
-    for pack in packs_to_install:
-        should_install, reason = should_install_content_pack(pack, marketplace_version, id_set,)
-        if not should_install:
-            packs_that_should_not_be_installed.add(f'{pack}: {reason}')
-        else:
-            packs_that_should_be_installed.add(pack)
-    if packs_that_should_not_be_installed:
-        logging.debug('The following packs should not be installed and therefore not collected: \n{} '.format(
-            '\n'.join(packs_that_should_not_be_installed)))
-
-    return packs_that_should_be_installed
-
-
-def is_documentation_changes_only(files_string: str) -> bool:
-    """
-
-    Args:
-        files_string: The modified files.
-
-    Returns: True is only documentation related files has been changed else False.
-
-    """
-    # Check if only README file in file string, if so, no need to create the servers.
-    files = [s for s in files_string.split('\n') if s]
-    documentation_changes_only = \
-        all(map(lambda s: s.endswith('.md') or s.endswith('.png') or s.endswith('.jpg') or s.endswith('.mp4'), files))
-    return documentation_changes_only
-
-
-def get_test_list_and_content_packs_to_install(files_string,
-                                               branch_name,
-                                               marketplace_version,
-                                               conf=deepcopy(CONF),
-                                               id_set=deepcopy(ID_SET)):
-    """Create a test list that should run"""
-    if marketplace_version == 'marketplacev2':
-        files_string = filter_modified_files_for_specific_marketplace_version(files_string, id_set, marketplace_version)
-        logging.debug(f'Files string after filter: {files_string}')
-
-    modified_files_instance = get_modified_files_for_testing(files_string)
-
-    modified_files_with_relevant_tests = modified_files_instance.modified_files
-    modified_tests_list = modified_files_instance.modified_tests
-    changed_common = modified_files_instance.changed_common_files
-    is_conf_json = modified_files_instance.is_conf_json
-    sample_tests = modified_files_instance.sample_tests
-    modified_packs = modified_files_instance.modified_metadata
-    is_reputations_json = modified_files_instance.is_reputations_json
-    is_indicator_json = modified_files_instance.is_indicator_json
-    is_landing_page_sections_json = modified_files_instance.is_landing_page_sections_json
-
-    all_modified_files_paths = set(
-        modified_files_with_relevant_tests + modified_tests_list + changed_common + sample_tests
-    )
-
-    from_version, to_version = get_from_version_and_to_version_bounderies(all_modified_files_paths,
-                                                                          id_set,
-                                                                          modified_packs=modified_packs,
-                                                                          )
-
-    # Check if only README file in file string, if so, no need to create the servers.
-    documentation_changes_only = is_documentation_changes_only(files_string)
-    create_filter_envs_file(from_version, to_version, documentation_changes_only=documentation_changes_only)
-
-    tests = set([])
-    packs_to_install = set([])
-
-    # Get packs and tests for changed scripts integration and playbooks
-    if modified_files_with_relevant_tests:
-        tests, packs_to_install = find_tests_and_content_packs_for_modified_files(modified_files_with_relevant_tests,
-                                                                                  conf, id_set)
-
-    # Adding a unique test for a json file.
-    if is_reputations_json:
-        tests.add('FormattingPerformance - Test')
-        tests.add('reputations.json Test')
-        tests.add('Indicators reputation-.json Test')
-
-    if is_indicator_json:
-        tests.add('Test IP Indicator Fields')
-
-    for file_path in modified_tests_list:
-        test = tools.collect_ids(file_path)
-        if test not in tests:
-            tests.add(test)
-
-    if is_conf_json:
-        tests = tests.union(get_test_from_conf(branch_name, conf))
-
-    if is_landing_page_sections_json:
-        packs_to_install |= get_packs_from_landing_page(branch_name)
-
-    if changed_common:
-        tests.add('TestCommonPython')
-
-    # get all modified packs - not just tests related
-    # TODO: need to move the logic of collecting packs of all items to be inside get_modified_files_for_testing
-    modified_packs = get_modified_packs(files_string)
-    if modified_packs:
-        packs_to_install = packs_to_install.union(modified_packs)
-
-    # Get packs of integrations corresponding to each test, as listed in conf.json
-    packs_of_tested_integrations = conf.get_packs_of_tested_integrations(tests, id_set)
-    packs_to_install = packs_to_install.union(packs_of_tested_integrations)
-
-    # Get packs that contains each of the collected tests
-    packs_of_collected_tests = get_content_pack_name_of_test(tests, id_set)
-    packs_to_install = packs_to_install.union(packs_of_collected_tests)
-
-    # All filtering out of packs should be done here
-    packs_to_install = filter_installed_packs(packs_to_install, marketplace_version, id_set)
-
-    # All filtering out of tests should be done here
-    tests = filter_tests(tests, id_set, modified_packs, marketplace_version, False)
-
-    if marketplace_version != 'marketplacev2':
-        if not tests or changed_common:
-            if not tests:
-                logging.info("No tests found running sanity check only.")
-            else:
-                logging.info("Changed one of the Common Server files, running sanity check too.")
-
-            logging.debug(f"Adding sanity tests: {SANITY_TESTS}")
-            tests.update(SANITY_TESTS)
-            logging.debug("Adding HelloWorld to tests as most of the sanity tests requires it.")
-            logging.debug(
-                "Adding Whois to packs to install as 'Sanity Test - Playbook with Unmockable Whois Integration' uses it"
-            )
-            packs_to_install.update(["HelloWorld", "Whois"])
-
-    # We add Base and DeveloperTools packs for every build
-    if tests or packs_to_install:
-        packs_to_install.update(["DeveloperTools", "Base"])
-
-    return tests, packs_to_install
-
-
-def get_all_packs_with_artifacts_for_specific_marketplace(id_set: dict, marketplace_version: str):
-    """Get all packs that have content artifacts in specific marketplace version:
-        Args:
-            id_set (dict): The id set object
-            marketplace_version (str): The marketplace version were the tests will run
-
-        Return:
-             set: set of packs
-        """
-    packs_to_install = set()
-    for artifacts in id_set.values():
-        # Ignore the Packs list in the ID set
-        if isinstance(artifacts, dict):
-            continue
-        for artifact_dict in artifacts:
-            for artifact_details in artifact_dict.values():
-                if artifact_details.get('marketplaces', []) == [marketplace_version]:
-                    packs_to_install.add(artifact_details.get('pack'))
-    return packs_to_install
-
-
-def get_test_playbooks_for_specific_marketplace(id_set: dict, marketplace_version: str):
-    """Get all test playbooks for specific marketplace version:
-        Args:
-            id_set (dict): The id set object
-            marketplace_version (str): The marketplace version were the tests will run
-
-        Return:
-             set: set test playbooks name
-        """
-    tests = set()
-    for test_obj in id_set.get('TestPlaybooks', []):
-        for test_id, test_data in test_obj.items():
-            if test_data.get('marketplaces', []) == [marketplace_version]:
-                tests.add(test_data.get('name'))
-    return tests
-
-
-def get_from_version_and_to_version_bounderies(all_modified_files_paths: set,
-                                               id_set: dict,
-                                               modified_packs: set = None) -> Tuple[str, str]:
-    """Computes the lowest from version of the modified files, the highest from version and the highest to version of
-    the modified files.
-    In case that max_from_version is higher than max to version - to version will be the the highest default.
-
-    Args:
-        all_modified_files_paths: All modified files
-        id_set: The content of the id.set_json
-        modified_packs: A set of modified pack names
-
-    Returns:
-        (string, string). The boundaries of the lowest from version (defaults to 0.0.0)
-         and highest to version (defaults to 99.99.99)
-    """
-    if all_modified_files_paths == {'Tests/Marketplace/landingPage_sections.json'}:
-        logging.debug('landingPage_sections.json is the only modified file, running only marketplace instances')
-        return '6.0.0', '99.99.99'
-    modified_packs = modified_packs if modified_packs else set([])
-    max_to_version = Version('0.0.0')
-    min_from_version = Version('99.99.99')
-    max_from_version = Version('0.0.0')
-
-    logging.info("\n\n Tests list:")
-    logging.info(modified_packs)
-    for pack_name in modified_packs:
-        pack_metadata_path = os.path.join(tools.pack_name_to_path(pack_name), constants.PACKS_PACK_META_FILE_NAME)
-        pack_metadata = tools.get_pack_metadata(pack_metadata_path)
-        from_version = pack_metadata.get('serverMinVersion')
-        to_version = pack_metadata.get('serverMaxVersion')
-        if from_version:
-            min_from_version = min(min_from_version, Version(from_version))
-            max_from_version = max(max_from_version, Version(from_version))
-        if to_version:
-            max_to_version = max(max_to_version, Version(to_version))
-
-    for artifacts in id_set.values():
-
-        # Ignore the Packs list in the ID set
-        if isinstance(artifacts, dict):
-            break
-
-        for artifact_dict in artifacts:
-            for artifact_details in artifact_dict.values():
-                if artifact_details.get('file_path') in all_modified_files_paths:
-                    from_version = artifact_details.get('fromversion')
-                    to_version = artifact_details.get('toversion')
-                    if from_version:
-                        min_from_version = min(min_from_version, Version(from_version))
-                        max_from_version = max(max_from_version, Version(from_version))
-                    if to_version:
-                        max_to_version = max(max_to_version, Version(to_version))
-
-    if str(max_to_version) == '0.0.0' or max_to_version < max_from_version:
-        max_to_version = Version('99.99.99')
-    if str(min_from_version) == '99.99.99':
-        min_from_version = Version('0.0.0')
-    logging.debug(f'modified files are {all_modified_files_paths}')
-    logging.debug(f'lowest from version found is {min_from_version}')
-    logging.debug(f'highest from version found is {max_from_version}')
-    logging.debug(f'highest to version found is {max_to_version}')
-    return str(min_from_version), str(max_to_version)
-
-
-def create_filter_envs_file(from_version: str, to_version: str, documentation_changes_only: bool = False):
-    """
-    Create a file containing all the envs we need to run for the CI
-    Args:
-        from_version: Server from_version
-        to_version: Server to_version
-        documentation_changes_only: If the build is for documentations changes only - no need to create instances.
-
-    """
-    envs_to_test = {
-        'Server Master': True,
-        'Server 6.8': is_runnable_in_server_version(from_version, '6.8', to_version),
-        'Server 6.6': is_runnable_in_server_version(from_version, '6.6', to_version),
-        'Server 6.5': is_runnable_in_server_version(from_version, '6.5', to_version),
-
-    }
-
-    if documentation_changes_only:
-        # No need to create the instances.
-        envs_to_test = {
-            'Server Master': False,
-            'Server 6.8': False,
-            'Server 6.6': False,
-            'Server 6.5': False,
-        }
-
-    logging.info("Creating filter_envs.json with the following envs: {}".format(envs_to_test))
-    with open(os.path.join(ARTIFACTS_FOLDER, 'filter_envs.json'), 'w') as filter_envs_file:
-        json.dump(envs_to_test, filter_envs_file)
-
-
-def get_list_of_files_in_the_pack(path_to_pack):
-    file_paths = []
-    for root, dirs, files in os.walk(path_to_pack):
-        for file in files:
-            file_paths.append(os.path.join(root, file))
-    return file_paths
-
-
-def changed_files_to_string(changed_files):
-    files_with_status = []
-
-    for file_path in changed_files:
-        file_with_status = f'M\t{file_path}'
-        files_with_status.append(file_with_status)
-
-    return '\n'.join(files_with_status)
-
-
-def create_test_file(is_nightly, skip_save=False, path_to_pack='', marketplace_version='xsoar', service_account=None):
-    """Create a file containing all the tests we need to run for the CI"""
-    if is_nightly:
-        if marketplace_version == 'marketplacev2':
-            # we are adding to the nightly on marketplacev2 few tests that are supported in both marketplacees
-            # see https://github.com/demisto/etc/issues/44350
-            tests = set(CONF.get_marketplacev2_tests())
-
-            # collect tests that are compatible only with marketplacev2
-            tests.update(get_test_playbooks_for_specific_marketplace(ID_SET, marketplace_version))
-
-            packs_to_install = CONF.get_packs_of_tested_integrations(tests, ID_SET)
-            packs_to_install.update(get_content_pack_name_of_test(tests, ID_SET))
-
-            # collect all packs and tests that are compatible only with marketplacev2
-            packs_to_install.update(get_all_packs_with_artifacts_for_specific_marketplace(ID_SET, marketplace_version))
-
-        else:
-            packs_to_install = (set(os.listdir(constants.PACKS_DIR)))
-            tests = set(CONF.get_test_playbook_ids())
-
-        packs_to_install = filter_installed_packs(packs_to_install, marketplace_version,
-                                                  deepcopy(ID_SET))
-        tests = filter_tests(tests, id_set=deepcopy(ID_SET), is_nightly=True,
-                             modified_packs=set(), marketplace_version=marketplace_version)
-        logging.info("Nightly - collected all tests that appear in conf.json and all packs from content repo that "
-                     "should be tested")
-    else:
-        branches = tools.run_command("git branch")
-        branch_name_reg = re.search(r"\* (.*)", branches)
-        if branch_name_reg:
-            branch_name = branch_name_reg.group(1)
-
-        logging.info("Getting changed files from the branch: {0}".format(branch_name))
-        if path_to_pack:
-            changed_files = get_list_of_files_in_the_pack(path_to_pack)
-            files_string = changed_files_to_string(changed_files)
-        elif os.environ.get("IFRA_ENV_TYPE") == 'Bucket-Upload':
-            last_upload_commit = get_last_commit_from_index(service_account)
-            current_commit = branch_name if branch_name != 'master' else 'origin/master'
-            files_string = tools.run_command(f'git diff --name-status {last_upload_commit}..{current_commit}')
-            logging.debug(f'Current commit: {current_commit}, Last upload commit: {last_upload_commit}')
-        elif branch_name != 'master':
-            files_string = tools.run_command("git diff --name-status origin/master...{0}".format(branch_name))
-            # Checks if the build is for contributor PR and if so add it's pack.
-            if os.getenv('CONTRIB_BRANCH'):
-                packs_diff = tools.run_command('git status -uall --porcelain -- Packs').replace('??', 'A')
-                files_string = '\n'.join([files_string, packs_diff])
-        else:
-            commit_string = tools.run_command("git log -n 2 --pretty='%H'")
-            logging.debug(f'commit string: {commit_string}')
-            commit_string = commit_string.replace("'", "")
-            last_commit, second_last_commit = commit_string.split()
-            files_string = tools.run_command(f'git diff --name-status {second_last_commit}...{last_commit}')
-
-        logging.debug(f'Files string: {files_string}')
-        files_string = '\n'.join(filter(lambda line: 'scripts/collect_tests' not in line, files_string.split('\n')))
-
-        tests, packs_to_install = get_test_list_and_content_packs_to_install(files_string, branch_name,
-                                                                             marketplace_version)
-
-    tests_string = '\n'.join(tests)
-    packs_to_install_string = '\n'.join(packs_to_install)
-
-    if not skip_save:
-        logging.info("Creating filter_file_old.txt")
-        with open(os.path.join(ARTIFACTS_FOLDER, 'filter_file_old.txt'), 'w') as filter_file:
-            filter_file.write(tests_string)
-        # content_packs_to_install_old.txt is not used in nightly build
-        logging.info("Creating content_packs_to_install_old.txt")
-        with open(os.path.join(ARTIFACTS_FOLDER, 'content_packs_to_install_old.txt'), 'w') as content_packs_to_install:
-            content_packs_to_install.write(packs_to_install_string)
-
-    if is_nightly:
-        logging.debug('Collected the following tests:\n{0}\n'.format(tests_string))
-        logging.debug('Collected the following packs to install:\n{0}\n'.format('\n'.join(packs_to_install)))
-
-    else:
-        if tests_string:
-            success_msg = 'Collected the following tests:\n{0}\n'.format(tests_string)
-            logging.success(success_msg)
-        else:
-            logging.error('Did not find tests to run')
-
-        if packs_to_install_string:
-            success_msg = 'Collected the following content packs to install:\n{0}\n'.format(packs_to_install_string)
-            logging.success(success_msg)
-        else:
-            logging.error('Did not find content packs to install')
-
-
-if __name__ == "__main__":
-    install_logging('Collect_Tests_And_Content_Packs.log', logger=logging)
-    logging.info("Starting creation of test filter file")
-
-    parser = argparse.ArgumentParser(description='Utility CircleCI usage')
-    parser.add_argument('-n', '--nightly', type=tools.str2bool, help='Is nightly or not')
-    parser.add_argument('-s', '--skip-save', type=tools.str2bool,
-                        help='Skipping saving the test filter file (good for simply doing validation)')
-    parser.add_argument('-p', '--changed_pack_path', type=str, help='A string representing the changed files')
-    parser.add_argument('-mp', '--marketplace', help='marketplace version.', default='xsoar')
-    parser.add_argument('--service_account', help="Path to gcloud service account")
-    options = parser.parse_args()
-
-    # Create test file based only on committed files
-    create_test_file(options.nightly, options.skip_save, options.changed_pack_path, options.marketplace,
-                     options.service_account)
-    if not _FAILED:
-        logging.info("Finished test configuration")
-        sys.exit(0)
-    else:
-        logging.error("Failed test configuration. See previous errors.")
-        sys.exit(1)
